requirements.txt
biopython>=1.73
plotly>=3.3.0
CAI>=1.0.3
